@@ -1,13 +1,13 @@
-package com.epam.webapp.command;
+package com.epam.webapp.command.client;
 
+import com.epam.webapp.command.Command;
+import com.epam.webapp.command.CommandResult;
 import com.epam.webapp.exception.ServiceException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public class ProgramCommand implements Command {
-
-     private static final String CLIENT_PROGRAM_PAGE = "WEB-INF/jsp/client/program.jsp";
 
      @Override
     public CommandResult execute(HttpServletRequest req, HttpServletResponse resp) throws ServiceException {
