@@ -1,5 +1,8 @@
-package com.epam.webapp.command;
+package com.epam.webapp.command.client;
 
+import com.epam.webapp.command.Command;
+import com.epam.webapp.command.CommandResult;
+import com.epam.webapp.constant.SessionConstant;
 import com.epam.webapp.entity.Program;
 import com.epam.webapp.exception.ServiceException;
 import com.epam.webapp.service.ProgramService;
@@ -9,24 +12,23 @@ import org.apache.logging.log4j.Logger;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
-import java.util.Enumeration;
 import java.util.Optional;
 
-public class ShowFinishedProgramPageCommand implements Command{
+public class ShowFinishedProgramCommand implements Command {
 
     private static final Logger LOGGER = LogManager.getLogger();
 
     private final ProgramService service = new ProgramService();
 
-    private static final String PROGRAM_PAGE_VALUE = "WEB-INF/jsp/client/program.jsp";
+    private static final String PROGRAM_PAGE = "WEB-INF/jsp/client/program.jsp";
+    private static final String PROGRAM_ID_PARAMETER_NAME = "programId";
 
     @Override
     public CommandResult execute(HttpServletRequest req, HttpServletResponse resp) throws ServiceException {
         HttpSession session = req.getSession();
-        long programId = Integer.parseInt((String) req.getParameter("programId"));
-        LOGGER.info("programId = {}",programId);
+        long programId = Integer.parseInt(req.getParameter(PROGRAM_ID_PARAMETER_NAME));
         Optional<Program> program = service.showFinishedProgram(programId);
-        session.setAttribute("program",program.get());
-        return CommandResult.forward(PROGRAM_PAGE_VALUE);
+        session.setAttribute(SessionConstant.CLIENT_PROGRAM,program.get());
+        return CommandResult.forward(PROGRAM_PAGE);
     }
 }
