@@ -1,9 +1,12 @@
-package com.epam.webapp.command;
+package com.epam.webapp.command.common;
 
+import com.epam.webapp.command.Command;
+import com.epam.webapp.command.CommandResult;
+import com.epam.webapp.constant.SessionConstant;
 import com.epam.webapp.entity.Client;
 import com.epam.webapp.entity.Trainer;
 import com.epam.webapp.exception.ServiceException;
-import com.epam.webapp.service.UserService;
+import com.epam.webapp.service.ClientService;
 import com.epam.webapp.service.TrainerService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -17,7 +20,7 @@ public class LoginCommand implements Command {
     private static final Logger LOGGER = LogManager.getLogger();
 
     private final TrainerService trainerService = new TrainerService();
-    private final UserService clientService = new UserService();
+    private final ClientService clientService = new ClientService();
 
     private static final String LOGIN_PARAMETER = "login";
     private static final String PASSWORD_PARAMETER = "password";
@@ -33,13 +36,13 @@ public class LoginCommand implements Command {
         Optional<Trainer> trainer = trainerService.login(login, password);
 
         if (client.isPresent()) {
-            req.getSession().setAttribute("user", client.get());
+            req.getSession().setAttribute(SessionConstant.USER, client.get());
             return CommandResult.redirect("controller?command=" + "showMainPage");
         } else if (trainer.isPresent()) {
-            req.getSession().setAttribute("user", trainer.get());
+            req.getSession().setAttribute(SessionConstant.USER, trainer.get());
             return CommandResult.redirect("controller?command=" + "showMainPage");
         } else {
-            req.getSession().setAttribute("errorMessage", "invalid credentials");
+            req.getSession().setAttribute(SessionConstant.ERROR_MESSAGE, "invalid credentials");
             return CommandResult.forward(LOGIN_PAGE);
         }
 
