@@ -1,10 +1,8 @@
 package com.epam.webapp.dao;
 
 import com.epam.webapp.connection.ProxyConnection;
-import com.epam.webapp.dao.impl.ClientDaoImpl;
-import com.epam.webapp.dao.impl.ExerciseDaoImpl;
-import com.epam.webapp.dao.impl.ProgramDaoImpl;
-import com.epam.webapp.dao.impl.TrainerDaoImpl;
+import com.epam.webapp.dao.impl.*;
+import com.epam.webapp.exception.DaoException;
 
 import java.sql.SQLException;
 
@@ -12,7 +10,7 @@ public class DaoHelper implements AutoCloseable {
     private final ProxyConnection connection;
 
     public DaoHelper(ProxyConnection connection) {
-       this.connection = connection;
+        this.connection = connection;
     }
 
     public ClientDaoImpl createClientDao() {
@@ -31,8 +29,28 @@ public class DaoHelper implements AutoCloseable {
         return new ExerciseDaoImpl(connection);
     }
 
+    public OrderDaoImpl createOrderDao() {
+        return new OrderDaoImpl(connection);
+    }
+
+    public void startTransaction() throws DaoException {
+        try {
+            connection.setAutoCommit(false);
+        } catch (SQLException e) {
+            throw new DaoException(e);
+        }
+    }
+
+    public void endTransaction() throws DaoException {
+        try {
+            connection.commit();
+        } catch (SQLException e) {
+            throw new DaoException(e);
+        }
+    }
+
     @Override
     public void close() throws SQLException {
-      connection.close();
+        connection.close();
     }
 }
