@@ -33,29 +33,42 @@ public class ProgramDaoImpl extends AbstractDao<Program> implements ProgramDao<P
                     """;
     private static final String SQL_SELECT_FIND_FINISHED_PROGRAM_BY_ID =
             "SELECT * FROM `program` WHERE id = ?";
+    private static final String SQL_SELECT_FIND_PROCESS_PROGRAM_BY_CLIENT_ID =
+            """
+                    SELECT * FROM program
+                    WHERE order_id = (SELECT id FROM `order` WHERE client_id = ?)
+                    AND state = 'process'
+                    """;
+    private static final String SQL_UPDATE_FEEDBACK =
+            """
+                    UPDATE `program`
+                    SET `feedback` = ?
+                    WHERE `id` = ?
+                    """;
+    private static final String SQL_UPDATE_STATE =
+            """
+                    UPDATE `program`
+                    SET `state` = ?
+                    WHERE `id` = ?
+                                      """;
 
     public ProgramDaoImpl(Connection connection) {
         super(connection, new ProgramRowMapper());
     }
 
     @Override
-    public String getTableName() {
-        return "program";
-    }
-
-    @Override
-    public Optional<Program> findProgramByOrderId(long orderId) throws DaoException {
+    public Optional<Program> findByOrderId(long orderId) throws DaoException {
         return executeForSingleResult(SQL_SELECT_FIND_PROGRAM_BY_ORDER_ID, orderId);
     }
 
     @Override
-    public Optional<Program> findProgramByClientId(long clientId) throws DaoException {
-        return executeForSingleResult(SQL_SELECT_FIND_PROGRAM_BY_CLIENT_ID,clientId);
+    public Optional<Program> findByClientId(long clientId) throws DaoException {
+        return executeForSingleResult(SQL_SELECT_FIND_PROGRAM_BY_CLIENT_ID, clientId);
     }
 
     @Override
-    public Optional<Program> findProgramByTrainerId(long trainerId) throws DaoException {
-        return executeForSingleResult(SQL_SELECT_FIND_PROGRAM_BY_TRAINER_ID,trainerId);
+    public Optional<Program> findByTrainerId(long trainerId) throws DaoException {
+        return executeForSingleResult(SQL_SELECT_FIND_PROGRAM_BY_TRAINER_ID, trainerId);
     }
 
     public List<Program> findAllPrograms() throws DaoException {
@@ -63,13 +76,33 @@ public class ProgramDaoImpl extends AbstractDao<Program> implements ProgramDao<P
     }
 
     @Override
-    public Optional<Program> findFinishedProgramById(long programId) throws DaoException {
-        return executeForSingleResult(SQL_SELECT_FIND_FINISHED_PROGRAM_BY_ID,programId);
+    public Optional<Program> findFinishedProgramByClientId(long clientId) throws DaoException {
+        return executeForSingleResult(SQL_SELECT_FIND_FINISHED_PROGRAM_BY_ID, clientId);
     }
 
     @Override
-    public List<Program> findClientFinishedPrograms(long clientId) throws DaoException {
-        return executeQuery(SQL_SELECT_FIND_CLIENT_FINISHED_PROGRAMS,clientId);
+    public Optional<Program> findProcessProgramByClientId(long clientId) throws DaoException {
+        return executeForSingleResult(SQL_SELECT_FIND_PROCESS_PROGRAM_BY_CLIENT_ID,clientId);
+    }
+
+    @Override
+    public List<Program> findFinishedClientPrograms(long clientId) throws DaoException {
+        return executeQuery(SQL_SELECT_FIND_CLIENT_FINISHED_PROGRAMS, clientId);
+    }
+
+    @Override
+    public void updateFeedback(long programId, String feedback) throws DaoException {
+        update(SQL_UPDATE_FEEDBACK, feedback, programId);
+    }
+
+    @Override
+    public void updateProgramState(long programId, ProgramState state) throws DaoException {
+        update(SQL_UPDATE_STATE, state.toString(), programId);
+    }
+
+    @Override
+    public String getTableName() {
+        return Program.TABLE;
     }
 
 }
