@@ -1,27 +1,18 @@
 package com.epam.webapp.entity;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 import java.io.Serializable;
 
-public abstract class Entity implements Serializable, Cloneable {
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public abstract class Entity implements Serializable{
 
     protected long id;
 
-    public final static String ID = "id";
-
-    public Entity() {
-
-    }
-
-    public Entity(long id) {
-        this.id = id;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
+    public static final String ID = "id";
 
 }
