@@ -1,8 +1,17 @@
 package com.epam.webapp.entity;
 
-import java.util.Objects;
+import lombok.*;
 
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class Exercise extends Entity {
+
+    public static final String TABLE = "exercise";
+
     private String name;
 
     public Exercise(long id, String name) {
@@ -10,33 +19,4 @@ public class Exercise extends Entity {
         this.name = name;
     }
 
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Exercise exercise = (Exercise) o;
-        return Objects.equals(name, exercise.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name);
-    }
-
-    @Override
-    public String toString() {
-        return new StringBuilder().append("Exercise{")
-                .append("id=").append(id)
-                .append(", name='").append(name).append('\'')
-                .append('}')
-                .toString();
-    }
 }
